requirements.txt
ctranslate2==3.22.0
transformers==4.42.0
sentencepiece==0.1.99
unbabel-comet==2.2.6
transquest==1.0.0
python-docx==1.1.0          # si vous lisez du .docx
PyYAML==6.0
